requirements.txt
regex~=2020.11.13
yara-python~=4.0.2
word2number-i18n~=1.2.0
spa2num~=0.0.5
python-dotenv~=0.15.0
pattern==3.6
spacy~=3.0.6
networkx~=2.6.2
click==7.1.2
scikit-learn==1.0
pandas==1.3.2